initializer/pyproject.toml
[tool.poetry]
name = "retail-initializer"
version = "0.1.0"
description = "Utility scripts for the retail backend project"
authors = ["Brendan Copp brendancopp1@gmail.com"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
boto3 = "^1.26.137"
python-dotenv = "^1.0.0"
pydantic = "^2.4.2"
typing-extensions = "^4.7.1"
pytz = "^2023.3"
tenacity = "^8.2.2"
mysql-connector-python = "^8.0.26"
google-cloud-storage = "^2.10.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"
isort = "^5.12.0"
pre-commit = "^3.3.3"
pytest-mock = "^3.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 